move common downhole models to shared location
